refactor : name 패키지 변경 -> Global
