+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -6,6 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import kitchenpos.global.domain.Name;
 
 
 @Entity
