+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -7,7 +7,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.menu.domain.Name;
+import kitchenpos.global.domain.Name;
 import kitchenpos.global.domain.Price;
 import org.springframework.util.ObjectUtils;
