+++ src/test/java/kitchenpos/global/NameTest.java
@@ -1,10 +1,10 @@
-package kitchenpos.menu.domain;
+package kitchenpos.global;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import kitchenpos.menu.domain.Name;
+import kitchenpos.global.domain.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
